[painless] extract CosseratBeamGeometry<TStrain> mixin (option B step 2)
Move all the geometry-only machinery out of CosseratGeometryMapping into
a
new non-BaseObject mixin CosseratBeamGeometry<TStrain> that derived
mappings
inherit alongside Multi2Mapping (Z1 inheritance model). This decouples
the
"section/frame topology" concern from the "SOFA mechanical states"
concern
and prepares the ground for Frames2StrainCosseratMapping to switch to a
simple Mapping<TIn,TOut> in step 3.

New files:
- src/Cosserat/mapping/CosseratBeamGeometry.h (~555 lines)
  Contains SectionInfo, FrameInfo, BeamTopology, and the
  CosseratBeamGeometry<TStrain> template class holding:
    - geometry state (m_section_properties, m_frame_properties, …)
    - Data members (d_curv_abs_section/_frames/_debug — stored here,
      registered by derived class via addData())
    - geometry helpers (updateGeometryInfo, initializeSectionProperties,
      initializeFrameProperties, updateTangExpSE3, …)
  Does NOT inherit from BaseObject to avoid the diamond with
Multi2Mapping.
- src/Cosserat/mapping/CosseratBeamGeometry.inl (~155 lines)
  Template implementations of the geometry helpers.

Refactored:
- CosseratGeometryMapping.{h,inl} reduced from 867+319 to 77+94 lines.
  Inherits from both Multi2Mapping<TIn1, TIn2, TOut> and
  CosseratBeamGeometry<TIn1>; only retains m_strain_state, m_rigid_base,
  m_frames pointers plus init()/doBaseCosseratInit()/initialization()
hooks.
  Data registration uses addData() since the InitData pattern doesn't
apply
  when the Data is owned by a mixin base.
- CosseratGeometryMapping.cpp: add explicit instantiation of
  CosseratBeamGeometry<Vec3/Vec6/Rigid3> for proper Windows DLL exports.
- CMakeLists.txt: register the new header/inl pair.

Renames applied during extraction (user-validated):
  m_frameProperties           → m_frame_properties
  m_indices_vectors           → m_frame_to_section_indices
  m_indices_vectors_draw      → m_frame_to_section_indices_draw
  m_beam_length_vectors       → m_section_length_vectors
  m_frames_length_vectors     → m_frame_distance_to_section_start
  get_related_beam_index_()   → getRelatedSectionIndex()
  set_related_beam_index_()   → setRelatedSectionIndex()
  getDistanceToNearestBeamNode() → getDistanceToSectionStart()
  setDistanceToNearestBeamNode() → setDistanceToSectionStart()
  related_beam_index_         → related_section_index_ (private)
  distance_to_nearest_beam_node → distance_to_section_start (private)

Consumers updated via sed (5 files):
  Strain2FramesCosseratMapping.{h,inl}
  Strain2FramesCosseratMapping_debug.inl
  Frames2StrainCosseratMapping.{h,inl}
